requirements.txt
Django
South
gunicorn
mcash-mapi-client